Fotos
- Rediseño, agregada la opción de cambio de foto de perfil en la base de datos

+++ modelo/fotos.php
@@ -3,17 +3,20 @@
 class Foto extends Conexion{
     private $id;
     private $url;
+    private $perfil;
 
-    public function __construct($id, $url){
+    public function __construct($id, $url, $perfil){
         $this->id = $id;
         $this->url = $url;
+        $this->perfil = $perfil;
         
         $con = parent::conectar();
         try{
             
-            $query = $con->prepare("INSERT INTO fotos SET direccion=:urll, idAnimal=:id");
+            $query = $con->prepare("INSERT INTO fotos SET direccion=:urll, idAnimal=:id, perfil=:perfil");
             $query->bindParam(':urll', $this->url);
             $query->bindParam(':id', $this->id);
+            $query->bindParam(':perfil', $this->perfil);
             $query->execute(); 
             if($query->errorCode() != "00000"){
                 throw new Exception("Error al insertar la dirección de la imagen en la base de datos");
@@ -31,6 +34,19 @@ class Foto extends Conexion{
             $query->execute();
             $result = $query->fetch();
             return $result[0];
+
+        }catch(Exception $e){
+            exit("ERROR: ".$e->getMessage());
+        }
+    }
+
+    public function fotoPerfil($id){
+        $con = parent::conectar();
+        try{
+            $query = $con->prepare("SELECT * FROM fotos WHERE idAnimal=:id AND perfil=1");
+            $query->bindParam(':id', $id);
+            $query->execute();
+            return $query;
 
         }catch(Exception $e){
             exit("ERROR: ".$e->getMessage());
@@ -56,20 +72,25 @@ class Foto extends Conexion{
 
             $url = "../../publico/images/".$dir;
 
-            $eliminarFoto = unlink($url);
+            if(is_file($url)){
 
-            if($eliminarFoto == true){
+                $eliminarFoto = unlink($url);
 
-                $query = $con->prepare("DELETE FROM fotos WHERE cod=:cod");
-                $query->bindParam(":cod", $cod);
-                $query->execute();
-                if($query->errorCode() != "00000"){
-                    throw new Exception("No ha sido posible eliminar la foto de la base de datos");
+                if($eliminarFoto == true){
+
+                    $query = $con->prepare("DELETE FROM fotos WHERE cod=:cod");
+                    $query->bindParam(":cod", $cod);
+                    $query->execute();
+                    if($query->errorCode() != "00000"){
+                        throw new Exception("No ha sido posible eliminar la foto de la base de datos");
+                    }
+    
+                }else{
+                    echo "0%%";
                 }
-
-            }else{
-                echo "0%%";
             }
+
+           
         }catch(Exception $e){
             exit("ERROR AL ELIMINAR FOTO: ".$e->getMessage());
         }
